requirements.txt
aiohttp==3.9.0b0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-config-file==0.6.0
click-plugins==1.1.1
configobj==5.0.8
dataclasses-json==0.6.2
diskcache==5.6.3
distro==1.8.0
FLAML==2.1.1
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
humanize==4.8.0
idna==3.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.334
langsmith==0.0.63
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
ngrok==0.12.0
numpy==1.26.1
oauthlib==3.2.2
openai==1.0.0b3
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pyautogen==0.2.0b2
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tenacity==8.2.3
termcolor==2.3.0
tiktoken==0.5.1
tqdm==4.66.1
twarc==2.14.0
tweepy==4.14.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
yarl==1.9.2
